+++ CHANGELOG.md
@@ -5,6 +5,12 @@ PrestaShop 1.6.x, 1.7.x and 8.x
 
 # Changelog
 
+## [3.7.3]
+### Fixed
+- Fix: Payment fail issue when multiple discounts are applied.
+- Fix: Apple Pay payment mismatch issue with `thecheckout` module.
+- Fix: Apple Pay terminal sorting issue.
+ 
 ## [3.7.2]
 ### Added
 - Add terminal logo for SEPA